deepshield/requirements.txt
# ─────────────────────────────────────────────
#  DeepShield KYC  –  Python Dependencies
# ─────────────────────────────────────────────

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9

# Computer Vision & Image Processing
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0

# Deep Learning (Deepfake Classifier)
torch==2.3.0
torchvision==0.18.0
timm==1.0.3                  # EfficientNet-B4 pretrained weights

# Face Detection & Recognition
mediapipe==0.10.14
insightface==0.7.3           # ArcFace face matching
onnxruntime==1.18.0

# Audio Processing
librosa==0.10.2
soundfile==0.12.1
pyaudio==0.2.14

# OCR & Document Processing
pytesseract==0.3.10
easyocr==1.7.1
pdf2image==1.17.0

# Signal Processing
scikit-learn==1.4.2

# Data & Validation
pydantic==2.7.1
pydantic-settings==2.3.0

# Async & Concurrency
anyio==4.4.0
httpx==0.27.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
loguru==0.7.2
python-dotenv==1.0.1
redis==5.0.4
celery==5.4.0

# Monitoring
prometheus-fastapi-instrumentator==6.1.0
